split mapnik_image.cpp
