+++ binding.gyp
@@ -13,6 +13,20 @@
         "src/mapnik_geometry.cpp",
         "src/mapnik_feature.cpp",
         "src/mapnik_image.cpp",
+        "src/mapnik_image_clear.cpp",
+        "src/mapnik_image_composite.cpp",
+        "src/mapnik_image_copy.cpp",
+        "src/mapnik_image_encode.cpp",
+        "src/mapnik_image_fill.cpp",
+        "src/mapnik_image_filter.cpp",
+        "src/mapnik_image_from_buffer.cpp",
+        "src/mapnik_image_from_bytes.cpp",
+        "src/mapnik_image_open.cpp",
+        "src/mapnik_image_premultiply.cpp",
+        "src/mapnik_image_resize.cpp",
+        "src/mapnik_image_save.cpp",
+        "src/mapnik_image_solid.cpp",
+        "src/mapnik_image_svg.cpp",
         "src/mapnik_image_view.cpp",
         "src/mapnik_grid.cpp",
         "src/mapnik_grid_view.cpp",
